Custom exception handling

+++ src/main/java/com/springboot/web/springbootfirstwebapp/controller/TodoController.java
@@ -67,6 +67,8 @@ public class TodoController {
 	
 	@RequestMapping(value="/delete-todo", method=RequestMethod.GET)
 	public String deleteTodo( @RequestParam int id){
+		if(id ==1)
+			throw new RuntimeException("chuj");
 		service.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
